add parseData
